command statements test fix
