+++ joern-cli/frontends/rubysrc2cpg/src/main/scala/io/joern/rubysrc2cpg/astcreation/AstForStatementsCreator.scala
@@ -1,5 +1,6 @@
 package io.joern.rubysrc2cpg.astcreation
 
+import better.files.File
 import io.joern.rubysrc2cpg.parser.RubyParser._
 import io.joern.rubysrc2cpg.passes.Defines
 import io.joern.x2cpg.Ast
@@ -10,6 +11,7 @@ import io.shiftleft.codepropertygraph.generated.nodes.{
   NewCall,
   NewControlStructure,
   NewIdentifier,
+  NewImport,
   NewLiteral
 }
 
@@ -174,31 +176,16 @@ trait AstForStatementsCreator { this: AstCreator =>
     methodNameAsIdentifierStack.push(methodIdentifierAsts.head)
     val argsAsts = astForArgumentsWithoutParenthesesContext(ctx.argumentsWithoutParentheses())
 
-    val callNodes = methodIdentifierAsts.head.nodes.filter(node => node.isInstanceOf[NewCall])
+    val callNodes = methodIdentifierAsts.head.nodes.collect { case x: NewCall => x }
     if (callNodes.size == 1) {
-      val callNode = callNodes.head.asInstanceOf[NewCall]
-      if (
-        callNode.name == "require" ||
-        callNode.name == "require_once" ||
-        callNode.name == "load"
-      ) {
-        val literalImports = argsAsts.head.nodes
-          .filter(node => node.isInstanceOf[NewLiteral])
-
-        if (literalImports.size == 1) {
-          val importedFile =
-            literalImports.head
-              .asInstanceOf[NewLiteral]
-              .code
-          println(s"AST to be created for imported file ${importedFile}")
-        } else {
-          println(
-            s"Cannot process import since it is determined on the fly. Just creating a call node for later processing"
-          )
-          Seq(callAst(callNode, argsAsts))
-        }
+      val callNode = callNodes.head
+      if (callNode.name == "require" || callNode.name == "load") {
+        resolveRequireOrLoadPath(argsAsts, callNode)
+      } else if (callNode.name == "require_relative") {
+        resolveRelativePath(filename, argsAsts, callNode)
+      } else {
+        Seq(callAst(callNode, argsAsts))
       }
-      Seq(callAst(callNode, argsAsts))
     } else {
       argsAsts
     }
@@ -226,5 +213,48 @@ trait AstForStatementsCreator { this: AstCreator =>
     case ctx: SuperCommandContext        => Seq(astForSuperCommand(ctx))
     case ctx: SimpleMethodCommandContext => astForSimpleMethodCommand(ctx)
     case ctx: MemberAccessCommandContext => astForMemberAccessCommand(ctx)
+  }
+
+  protected def resolveRequireOrLoadPath(argsAst: Seq[Ast], callNode: NewCall): Seq[Ast] = {
+    val importedNode = argsAst.head.nodes.collect { case x: NewLiteral => x }
+    if (importedNode.size == 1) {
+      val node      = importedNode.head
+      val pathValue = node.code.replaceAll("'", "").replaceAll("\"", "")
+      val result = pathValue match {
+        case path if File(path).exists =>
+          path
+        case path if File(s"$path.rb").exists =>
+          s"${path}.rb"
+        case _ =>
+          pathValue
+      }
+      packageStack.append(result)
+      astForImportNode(node.code)
+    } else {
+      Seq(callAst(callNode, argsAst))
+    }
+  }
+
+  protected def resolveRelativePath(currentFile: String, argsAst: Seq[Ast], callNode: NewCall): Seq[Ast] = {
+    val importedNode = argsAst.head.nodes.collect { case x: NewLiteral => x }
+    if (importedNode.size == 1) {
+      val node        = importedNode.head
+      val pathValue   = node.code.replaceAll("'", "").replaceAll("\"", "")
+      val updatedPath = if (pathValue.endsWith(".rb")) pathValue else s"$pathValue.rb"
+
+      val currentDirectory = File(currentFile).parent
+      val file             = File(currentDirectory, updatedPath)
+      packageStack.append(file.pathAsString)
+      astForImportNode(node.code)
+    } else {
+      Seq(callAst(callNode, argsAst))
+    }
+  }
+
+  protected def astForImportNode(code: String): Seq[Ast] = {
+    // fully implemented later
+    val importNode = NewImport()
+      .code(code)
+    Seq(Ast(importNode))
   }
 }
